Problems on older machine addressed.
Former-commit-id: a70ade915905f9ba6d3941243c7de9768e361397
Former-commit-id: 19d5aa6255cc72d552920590d323eb56a96f506f

+++ App/src/DataFile/gdf2.cpp
@@ -146,7 +146,7 @@ GDF2::GDF2(const string& filePath)
 	vh.sensorInfo = new char[getChannelCount()][20];
 	readFile(*vh.sensorInfo, 20*getChannelCount());
 
-	assert(ftell(file) == 256 + 256*getChannelCount());
+	assert(ftell(file) == static_cast<int>(256 + 256*getChannelCount()));
 
 	// Initialize other members.
 	samplesRecorded = vh.samplesPerRecord[0]*fh.numberOfDataRecords;
